requirements.txt
aiogram~=3.4.1
python-dotenv~=1.0.1
requests~=2.31.0
beautifulsoup4~=4.12.3